a bit of debugging
